fix: split context env-write wrappers out of def_mapping.rs to satisfy the 2000-line ceiling
